Started refactor PhoneNumberInput
