+++ client/src/components/form-components/Input/Input.jsx
@@ -1,5 +1,10 @@
 import cl from './index.module.scss';
 
-export function Input({ name, variant, register, ...rest }) {
-  return <input className={`${cl.input} ${cl[variant]}`} {...register(name)} {...rest} />;
+export function Input({ labelText, name, variant, register, ...rest }) {
+  return (
+    <>
+      <label htmlFor={name}>{labelText}</label>
+      <input id={name} className={`${cl.input} ${cl[variant]}`} {...register(name)} {...rest} />
+    </>
+  );
 }
