+++ client/src/components/form-components/index.js
@@ -3,5 +3,5 @@ export { FormGroup } from './FormGroup/FormGroup.jsx';
 export { Input } from './Input/Input.jsx';
 export { Select } from './Select/Select.jsx';
 export { ErrorMessage } from './Error/ErrorMessage.jsx';
-export { PhoneNumberInput } from './Input/PhoneNumberInput.jsx';
+export { PhoneNumberInput } from './Input/PhoneNumberInput/PhoneNumberInput.jsx';
 export { Textarea } from './Textarea/Textarea.jsx';
